Change cctools_version_debug to accept 64 integer.

+++ dttools/src/cctools.c
@@ -18,7 +18,7 @@ void cctools_version_print (FILE *stream, const char *cmd)
 	fprintf(stream, "\tConfiguration: %s\n", CCTOOLS_CONFIGURE_ARGUMENTS);
 }
 
-void cctools_version_debug (int type, const char *cmd)
+void cctools_version_debug (uint64_t type, const char *cmd)
 {
 	debug(type, "%s version %s (released %s)", cmd, CCTOOLS_VERSION, CCTOOLS_RELEASE_DATE);
 	debug(type, "Built by %s@%s on %s at %s", BUILD_USER, BUILD_HOST, __DATE__, __TIME__);
